CLUE-646: move pure drawing geometry where the summarizer can reach it
